+++ CHANGES.rst
@@ -7,6 +7,7 @@ Version 0.9.1 (UNRELEASED)
 - Adds ``prune`` command to delete all intermediate files of a given workflow.
 - Fixes ``list`` command to correctly list workflows when sorting them by their run number or by the size of their workspace.
 - Fixes ``du`` command help message typo.
+- Fixes ``validation --environments`` command to correctly handle fully qualified image names.
 
 Version 0.9.0 (2023-01-26)
 --------------------------